requirements.txt
Pillow>=10.0.0
hypothesis>=6.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
